arch-wrpc (and doc): finally fix the clock class
This sets the clock class in arch-wrpc in a way more aligned to the
standard. A grandmaster is 6 if locked to atomic or gps, but
52 if it is designated grandmaster but doesn't lock to a primary source
(this is was we poor developers experience).  A "master" is 187,
the old default, and a slave-only remains 255 as it has always been.

This also removes the ability to stop the program while it waits
for the pll to lock. It can only succeed or timeout.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>