build: avoid libraries: arch and protocols include OBJ-y instead
This simplifies the build system, by relying in weak symbols instead
of the link order. We need weak symbols anyways for configuration, so
we can get rid of the libraries.  The size of binaries returned at
build time is increased, because arch/ and lib/ stuff is not part of
ppsi.o (as it actually should).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>