wr-servo: kill a needless function
timeint_to_wr was actually copying data to a local uninitialized
variable of the same type as the input one.  The uninitialized
structure generates warnings about unitialized fields (at least on
newer compilers than the one we use on wr-switch).

Instead of clearing the local variable, just use structure assignment
and get rid of the function.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>