white rabbit: use wr_operations, not aliases
This is a massive change, but one that we needed since time 0.
Aliases were a bad choice, while use of an operation structure shows
better the symmetry between the two white rabbit implementations.
But, mainly, maintainance is easier.

I really needed this because I need to change one prototype, to
add support for wrpc acting as grandmaster.

There are still some asymmetries to be fixed, but the thing is
acceptable at this point. While I was at it, I removed some
WR_DSPORT(ppi) using a local pointer "wrp".

The code after this commit is shorter by a dozen bytes or so,
both on wrpc and wrs.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>