white rabbit: fold enable_timing_output in operations
This is the last bit of the overhaul made by the previous commit,
and has no technical effect. (In a perfect world, the previous commit
would have been one per function, to easily find errors, but I had not
enough time; this is different because it was a two-level call).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>