lib/conf.c: fix ARG_NONE, factorize code
Every case of the switch was colling the cfg_handler out of the
switch construct, avery time with the same arguments.
The only case not calling the config function was case ARG_NONE.
But this doesn't make sense, because if a configuration option
does not do anything, then it should not be there, even without
arguments.
So the call to config function is now made from every case, even
ARG_NONE, then it should not stay inside the switch.