bug workaround for slave states: don't refuse a new master
This works around a design problem that must be checked carefully.
The "is from current parent" flags set at receive time reveales bad,
as we should only check when needed.

So by now just say it is the current parent if we don't have a current
parent yet (i.e., we are not in error, even if it is not from current
parent).

Also, forget about current parent and foreign masters when the servo
is initialized (e.g., at link-up on the wr node).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>