Kconfig: added configurations; fixed Makefiles accordingly
This adds the needed configuration items: architectures, extension,
cross_compile and wrpc_root.  So, what we passed on environment and
command line is now in the configuration file.

This changes the priorities of such options, so the Makefile is
changed to force $(CC), $(LD) and $(CROSS_COMPILE) down to subprojects
(i.e. pp_printf and arch-wrs/mini-rpc).  Also, we need to remove
quotes from .config items, and this is ugly linesin ./Makefile

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>