wrpc: allow a build without included "ptpdump"
This is needed for the wrpc-sw release, as Etherbone won't fit with
a full ppsi in the default RAM size for SPEC cards.

ptpdump support will soon become optional in ppsi, through Kconfig,
and a bigger RAM size will become the default for SPEC images as well.
But currently we need this hack.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>