arch-sim: fields added to struct sim_ppi_arch_data
These are namely Data Sets, runtime options, servo and TimeProperties.
We need them in every instance because we have two ppi (master and slave)
acting like two different machines, so they need different copies of
these variables. We just put the per-instance stuff in the arch_data.
Then we'll change the pointers in the ppg directly in the net operations,
according to our needs.