arch-sim: added per-arch config options
startup is changed, because we need to set the initial time for the
master before we set the initial offset from master fo the slave.
In this way the diagnostics cannot be turned on to print out the
config for the master. This is not a problem, because the master
it's not meant to be configurable