Kconfig: fix to displayed version number
KERNELVERSION variable from Linux was still used.
It was not initialized so the displayed version in the headers
of make gconfig and make menuconfig was missingor wrong.
It just said (null) or KERNELVERSION.
KERNELVERSION has been changed to VERSION all over the kconfig
scripts. Makefile.kconfig exports this variable so it can be
used by config scripts with getenv().
Now it prints the ppsi version with the present commit, like
diplayed from
	git describe --always --dirty