std servo: no running avg and outlier removal on ofm
There is no need to remove outliers and to make a running
average on ofm. Indeed we know that ofm = t2 - t1 -mpd.
We are already removing outliers and averaging on mpd.
So, if after mpd has been "cleaned", ofm still shows outliers
or irregularities they are for sure due to t1 and t2. So they
are errors coming from the clocks and they have not to be
ignored. Instead they must be corrected ASAP, and to do that
the software must be able to see them instantly. So no
running avg nor outlier removal has to be performed on ofm values.