wrpc bugfix: time-wrpc/ objects are now in OBJ-y
Commit 84df8890 changed the build order, avoiding libraries.
Back then I fogot to apply the same change to time-wrpc, and didn't
notice because ./MAKEALL worked fine.  We never linkerd the final
ppsi.elf for WRPC since then.

Now I'm updating the ppsi submodule in the wrpc-sw repository, and
this must be fixed to avoid the missing symbols in the final link.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>