arch-sim: new simulator diagnostics and config
diagnostics:

For testing purposes we can't just read the ofm value
printed out by the slave, because that's only the offset
perceived by the slave and can be wrong. We need instead to
print out the offset obtained subtracting the real time of
the master from the real time of the slave.
To print out the ofm we use the "ext" flag of pp_diag.
The ofm is printed only when the slave gets a
DelayResp message from the master.
A new tool to strip ofm out of simulator log is provided

config:

now the max number of simulated ptp iterations can be configured.
previously one could set the number of seconds to simulate.
this is not possible anymore.