whiterabbit: remove unused assignments to wrPortState
The wrPortState variable was assigned-to in various places: it was
"almost" mirroring the PTP state, but not every WR state was actually
setting it.  In fact, it was only used for calibration sub-states, so
remove all other assignment.

Often, the code did:

   wp->wrMode = NON_WR;
   wp->wrPortState = WR_PORT_IDLE;

but only wrMode was ever checked, so WR_PORT_IDLE (was: WRS_IDLE) is
redundant and can be removed altogether.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>