whiterabbit: implement retry on timeout
This commits adds retry in most wr states. This allows to deal with
lost frames. There a few missing items, and a few rough situation.
The code itself shows a number of repetitions, which I dislike, but
over time we'll fix it properly.

By now this should be enough to deal with transient data loss
during calibration, that happens at some users' site.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>