general: rename freq_ppm to freq_ppb (and fix arch-bare)
All adjustments are passed as part-per-billion, so after getting mad
in understanding what freq_ppm was, I renamed all of them to freq_ppb.

While making changes I verified this is what it is. Thus I fixed the comment
in ppsi.h (that was wrong by a factor of 64) and the adjustment code in
bare-time.c that made the same error (I admit I don't test arch-bare often).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>