wrs: drive the PLL frequency when non-WR
This adds the "adjust_frequecy" method to time-wrs. The WR servo is
not using this method, which is only called by the normal PTP servo.

With this new method, a non-WR slave is able to trim frequency,
whereas before this commit we were only able to jump our time.  The
code move straight the FPGA register, by having some hardware-specific
constants explicit in the code. Thus, it is not very beautiful, but
it works fine on my systems.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>