general: peer and peer_vid are ppi-wide, not channel-wide
With UDP we have two channels but we talk with the same peers.
Moreover, we reply with a general message to an event message, so
having two peers was plain wrong.

I'm well aware this commit should happen before vlans are introduced,
but I don't want to deal with all the conflicts in moving it back.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>