arch-wrs: added shmem code from libwr
This is respinning a previous commit, based on the new shmem in libwr
after the feedback and patches by Adam. More headers are inserted,
as they are included by the ones we use.

hal_shmem.h has one 1-line modification to account for the different
include path in ppsi, as opposed to wr-switch-sw.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>