wrs-time: change Unix time too
time_operations->set() is rarely called for the white rabbit switch.
Besides the special case with a NULL time pointer, it is only called
by the normal servo, before WR-Mode is activated on the relevant link.
This happens once only.

In this situation, we should set Unix time, in addition to changing
the WR timers. The code, as laid out would change Unix time upon
repeated invocation of this function, but not the first time. Thus,
we risked keeping a wrong system time even if WR synced to the
external world.  This means syslog messages would be stamped at
January 1970.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>