general: move buffers our of pp_instance, and make them smaller
In order to fit all pp_instance items withing WRS shared memory,
we need to make them a little smaller.  Thus, the tx and rx buffers
are now allocated separately.  And, while I am at it, I make them
smaller, because 128 bytes are more than enough for PTP with
extensions (all frames are under 100 bytes). Given we are always
showr of memory in wrpc-sw, this is benefical.

As a side effect, clean up some out-of-memory exit paths, and fix a
use-before-check buglet in arch-unix (no, we never go out of memory,
but the check was wrong).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>