libwr: add new fields to hal_port_state
Function hal_port_get_exported_state, fills structure hal_port_state with data
that are not available in SHM. It just assigns defines.
These defines are private to HAL, and assignment should be done in HAL.

Similar problem is for wrpc. There is a wrapper function that pretends to be
HAL. All HAL info should be filled in that function.

Added:
--uint32_t clock_period;
--uint32_t t2_phase_transition;
--uint32_t t4_phase_transition;

Signed-off-by: Adam Wujek <adam.wujek@cern.ch>