arch-wrs: fixes for vlan
The way vlan frames are received is different on wrs than in normal
Linux. Likely my own driver (for wrs hardware) is not doing the right
thing.

So, when we receive tagged frames, they reach us as-is (not as the
payload only, with "aux" data on a side). This commit considers this
case.

The commit also fixes a pair of other details -- mostly, the amount of
"short packet of 0 bytes, ignored", by returning -2 to mean "drop with
no error".

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>

SQUASH fix for ARCH_WRS

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>