wrs-socket: bugfix: wait for late tx stamps
When the wrs has high data traffic, the tx stamp can be delayed more than
the time it takes for software to poll for it.

What happened is that we got back -EAGAIN, but when collecting the
next stamp we got the previous one, and sync goes nuts ever after.

This uses poll() to wait a little for the timestamp.

Moreover, we check the frame we get back, so we can discard it in case
some unsync between tx frames and timestamps happen -- checking frame
contents is the suggested policy anyways, but I've been lazy in this,
when I drafted this wrs support.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>