state-master: don't send f-up if sync stamp is wrong
This is bsased on Greg's original commit, which is in wrs-4.1.2-fixes,
that we developed and released in the wr-switch-sw maintainance for the
4.1 series.

What we do here is return error for sending if the timestamp is not
correctly retrieved; the error is already handled in the sync+f.up
pair, so this is enough for f.up to not be sent.

The commit also changes state-master, in that we need to restart the
timeout for sync (and annunce) even if one even of sync (announce,
resp) fails.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>