arch-wrs: pad frames sent with VLAN tag
The switch is padding short frames sent from the CPU in the nic core.
When the frame is then being sent to an "access" port, the tag is
removed, and thus the frame gets shorter.  The net effect is that
sync/f-up/delay-rep were sent short, and discarded by the recipient.

Greg could move padding to the endpoint, but this extra work is not
really needed, as nobody else is expected to get frames from the CPU
over the SFP ports. So we chose to fix transmission in software and
avoid to change the gateware.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>