white-rabbit: change the wr-servo string being reported
When the servo is tracking, it oscillates between two states:
track_phase and wait_sync_idle.  Users should better not see
the "wait_sync_idle", becuase it look problematic but it is not
(earlier, the state was entered in the same way but not shown).

We are going to change the thing, and speed-up the servo as a side
effect, but meanwhile we'd better do this so ppsi's master can be
picked as the wrpc-sw submodule

Please note that state reported as number and string via SNMP may be
inconsistent.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>