send functions: return -1 on error and 1 on stamping error (stamp invalid)
This is the first step to fix a problem introduced by commit 79ffd673
(a fix to release 4.1.2 forward ported to master).

When we get a bad timestamp in sending sync, we should not return "error"
but a special value for "no timestamp". Otherwise the caller goes to
"faulty" state and can't sync any more (this situation loops over).

Also, check for timestamp errors only for event message, or we break
out of the messaging loop during wr synchronization (when we get bad
timestamps for frames whose timestamp we don't care about).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>