protocols: handle PP_SEND_NO_STAMP specially
This is the result of an audit of all send functions. Most are not
affected, because they send on the "general" channel. But when event
messages are sent, an error of SEND_NO_STAMP must not be fatal.
In WR, for example, after a sync the stamping engine reports invalid
stamps for a second.

I chose not to change the return type of the sending function and have
the compiler help me, because my mates are working on peer-delay and I'd
better not make the merge more difficult for us.

Unfortunately, my wrs is still not syncing with this patch, as it remains
stuck in SYNC_NSEC (in the state machine within the WR servo). See
next commit....

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>