wr-servo: move WAIT_SYNC_IDLE state to a flag
This temporary wr-servo state, with the hack of next_state is not
really clean. I took a while to understand it, and this tries to make
it clearer for my fellow developer.

When we change state by commanding hardware, we must wait for hw to be
done with the change, and this is better a flag than another state.

While making this clean, the commit stresses the fact that we loose
one wr-servo iteration every two of them.  Unfortunately, the trivial
fix I tried some time ago doesn't actually work (the slave won't
sync), so the behaviour remains unchanged by this commit, but the
problem is not clearly visible in code and comments, so we may fix it
"the real way" later on.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>