whiterabbit: arch-wrs and arch-wrpc can't be build without WR support
This was a trial of mine, but it really has never been working (at least
on wrs -- I think nobody ever tried a non-wr code base on wrpc).

Too much of the timetamping, servo, pll, is dependent on WR being
available, this academic idea doesn't make much sense, so let's remove
a number of conditionals (and associated potential bugs).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>